Refactored how we record playback history to avoid scoping issue
