+++ infra/audioplayer/impl/src/androidMain/kotlin/app/campfire/audioplayer/impl/AudioPlayerService.kt
@@ -55,6 +55,7 @@ interface AudioPlayerComponent {
   val audioPlayerHolder: AudioPlayerHolder // AppScope
   val playbackSettings: PlaybackSettings // AppScope
   val activityIntentProvider: ActivityIntentProvider // AppScope
+  val exoPlayerFactory: ExoPlayerAudioPlayer.Factory // AppScope
 }
 
 @ContributesTo(UserScope::class)
@@ -62,7 +63,6 @@ interface AudioPlayerUserComponent {
   val mediaTree: MediaTree
   val sessionsRepository: SessionsRepository
   val playbackSessionManager: PlaybackSessionManager
-  val exoPlayerFactory: ExoPlayerAudioPlayer.Factory
 }
 
 @SuppressLint("UnsafeOptInUsageError")
@@ -85,16 +85,9 @@ class AudioPlayerService : MediaLibraryService() {
     super.onCreate()
     bark(LogPriority.INFO) { "AudioPlayerService::onCreate()" }
 
-    // Since we need to inject the exoplayer factory from UserScope,
-    // let's make sure its available, and block until it is.
-    val userComponent = runBlocking {
-      ComponentHolder.subscribe<AudioPlayerUserComponent>()
-        .first()
-    }
-
     // Create ExoPlayer instance and MediaSession instance that encapsulates the background
     // playback on Android.
-    player = userComponent.exoPlayerFactory.create(this)
+    player = component.exoPlayerFactory.create(this)
     // Use remoteControlPlayer for MediaSession so remote control commands (Bluetooth, car stereo)
     // can be intercepted and handled based on user settings, while in-app UI uses the direct player.
     session = MediaLibrarySession.Builder(this, player.player, MediaSessionCallback())
